Added utility object Dither Obstacle, to be added if the designer wants an object close to camera to dither
Instructions: place the dither object as a child of the object you want to disappear when the camera is close and sibling of the mesh. Tweak the values of min and max distance to get desired results. Be wary that this probably has a performance cost so do not abuse it.